requirements.txt
discord.py
tweepy
motor
dnspython
discord-components
pillow
emoji
psutil
